[SMALLFIX] Also move IndexedSet to collections
