+++ common/src/main/java/tachyon/collections/IndexedSet.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon;
+package tachyon.collections;
 
 import java.util.ArrayList;
 import java.util.HashMap;
